requirements.txt
playwright==1.40.0
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.0